+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Support handling of empty configuration responses (no bundle data) from qbee
+
 ## [0.4.0] - 2023-09-07
 
 ### Added